# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Transform Church Leadership Training Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    put:
      operationId: upsertMe
      tags: [users]
      summary: Create or update current user (called on first sign-in)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: Upserted user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Dashboard ───────────────────────────────────────────────────────────
...
  /tracks:
    get:
      operationId: listTracks
      tags: [tracks]
      summary: List all training tracks
      responses:
        "200":
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
    post:
      operationId: createTrack
      tags: [tracks]
      summary: Create a new track (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackInput"
      responses:
        "201":
          description: Created track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
  /tracks/{trackId}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get a single track with its modules
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Track with modules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackWithModules"
        "404":
          description: Not found
    patch:
      operationId: updateTrack
      tags: [tracks]
      summary: Update a track (admin only)
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackUpdate"
      responses:
        "200":
          description: Updated track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
    delete:
      operationId: deleteTrack
      tags: [tracks]
      summary: Delete a track (admin only)
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Modules ─────────────────────────────────────────────────────────────
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List videos (optionally filter by moduleId)
      parameters:
        - name: moduleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Create a new video (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Created video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /videos/{videoId}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
    patch:
      operationId: updateVideo
      tags: [videos]
      summary: Update a video (admin only)
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoUpdate"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video (admin only)
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Watch History ────────────────────────────────────────────────────────
...
  /modules/{moduleId}/quiz:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz questions for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestion"
    post:
      operationId: createQuizQuestion
      tags: [quizzes]
      summary: Add a quiz question to a module (admin only)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizQuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestion"
  /modules/{moduleId}/quiz/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers and get result
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /modules/{moduleId}/quiz/result:
    get:
      operationId: getQuizResult
      tags: [quizzes]
      summary: Get current user's quiz result for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz result (or null if not taken)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
        "404":
          description: Not taken yet
  /quizzes/questions/{questionId}:
    delete:
      operationId: deleteQuizQuestion
      tags: [quizzes]
      summary: Delete a quiz question (admin only)
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Assignments ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, admin]
        createdAt:
          type: string
      required: [id, clerkId, firstName, lastName, email, role, createdAt]
    UserInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
      required: [firstName, lastName, email]
...
    Track:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, createdAt]
    TrackInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required: [name]
    TrackUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    TrackWithModules:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        modules:
          type: array
          items:
            $ref: "#/components/schemas/Module"
      required: [id, name, createdAt, modules]
    Module:
      type: object
      properties:
        id:
          type: integer
        trackId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
      required: [id, trackId, title, order, createdAt]
...
    Video:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
        order:
          type: integer
        createdAt:
          type: string
      required: [id, moduleId, title, url, order, createdAt]
    VideoInput:
      type: object
      properties:
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
        order:
          type: integer
      required: [moduleId, title, url, order]
    VideoUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
        order:
          type: integer
...
    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        order:
          type: integer
      required: [id, moduleId, questionText, options, correctIndex, order]
    QuizQuestionInput:
      type: object
      properties:
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        order:
          type: integer
      required: [questionText, options, correctIndex, order]
    QuizSubmission:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              selectedIndex:
                type: integer
            required: [questionId, selectedIndex]
      required: [answers]
    QuizResult:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        moduleId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        passed:
          type: boolean
        takenAt:
          type: string
      required: [id, userId, moduleId, score, totalQuestions, passed, takenAt]
...