# lib/api-zod/src/generated/types/trackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Transform Church Leadership Training Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TrackInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}